refactor allocation list to be fully shared in all views
