+++ frontend/src/components/allocation_list.js
